+++ kaldiio/__init__.py
@@ -6,10 +6,10 @@ from kaldiio.matio import load_mat
 from kaldiio.matio import load_scp
 from kaldiio.matio import save_ark
 from kaldiio.matio import save_mat
-from kaldiio.highlevel import parse_specifier
 from kaldiio.highlevel import ReadHelper
 from kaldiio.highlevel import WriteHelper
 from kaldiio.utils import open_like_kaldi
+from kaldiio.utils import parse_specifier
 from kaldiio.wavio import load_wav_scp
 
 try:
