+++ kaldiio/utils.py
@@ -1,3 +1,4 @@
+from collections import OrderedDict
 from contextlib import contextmanager
 import io
 from io import TextIOBase
@@ -191,3 +192,55 @@ class MultiFileDescriptor(object):
             return string + string2
         else:
             return string
+
+
+def parse_specifier(specifier):
+    """A utility to parse "specifier"
+
+    Args:
+        specifier (str):
+    Returns:
+        parsed_dict (OrderedDict):
+            Like {'ark': 'file.ark', 'scp': 'file.scp'}
+
+
+    >>> d = parse_specifier('ark,t,scp:file.ark,file.scp')
+    >>> print(d['ark,t'])
+    file.ark
+
+    """
+    if not isinstance(specifier, str):
+        raise TypeError(
+            'Argument must be str, but got {}'.format(type(specifier)))
+    sp = specifier.split(':', 1)
+    if len(sp) != 2:
+        if ':' not in specifier:
+            raise ValueError('The output file must be specified with '
+                             'kaldi-specifier style,'
+                             ' e.g. ark,scp:out.ark,out.scp, but you gave as '
+                             '{}'.format(specifier))
+
+    types, files = sp
+    types = list((map(lambda x: x.strip(), types.split(','))))
+    files = list((map(lambda x: x.strip(), files.split(','))))
+    for x in set(types):
+        if types.count(x) > 1:
+            raise ValueError('{} is duplicated.'.format(x))
+
+    supported = [{'ark'}, {'scp'}, {'ark', 'scp'},
+                 {'ark', 't'}, {'scp', 'ark', 't'}]
+    if set(types) not in supported:
+        raise ValueError(
+            'Invalid type: {}, must be one of {}'.format(types, supported))
+
+    if 't' in types:
+        types.remove('t')
+        types[types.index('ark')] = 'ark,t'
+
+    if len(types) != len(files):
+        raise ValueError(
+            'The number of file types need to match with the file names: '
+            '{} != {}, you gave as {}'.format(len(types), len(files),
+                                              specifier))
+
+    return OrderedDict(zip(types, files))
